pyproject.toml
[project]
name = "nk-risk-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "accelerate>=1.13.0",
    "anthropic>=0.86.0",
    "bitsandbytes>=0.49.2",
    "datasets>=4.8.4",
    "db-dtypes>=1.5.0",
    "google-cloud-bigquery>=3.40.1",
    "gradio>=6.10.0",
    "pandas>=2.3.3",
    "peft>=0.18.1",
    "python-dotenv>=1.2.2",
    "torch>=2.11.0",
    "transformers>=5.3.0",
    "trl>=0.29.1",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
